api_users/requirements.txt
flask
pymongo
pyzmq
jsonschema
flasgger